+++ Makefile
@@ -2,7 +2,7 @@ CFLAGS= -g -Wextra -Wall -Werror
 LIBS	:= -lft -L./lib/libft
 NAME = minishell
 OBJECTS = src/main.o src/ft_split_quote.o src/minishell_utils.o src/sig_center.o src/sig_meta.o \
-	src/path_parsing.o
+	src/path_parsing.o	src/parent_child_process.o
 DEPS = -I./include -I./lib/libft
 READLINE_LIB = -L/usr/local/lib -I/usr/local/include -lreadline
 INCLUDES = include/minishell.h
